requirements.txt
Django==1.6.10
South==1.0.2
argparse==1.2.1
distribute==0.6.24
django-filter==0.9.2
django-filters==0.1.0
django-tastypie==0.12.1
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
wsgiref==0.1.2
